added sqlite3 db to setup

+++ setup.py
@@ -1,5 +1,6 @@
 import os
 import json
+import sqlite3
 
 #installs requirements, in case u don't have venv
 def install_requirements():
@@ -12,16 +13,16 @@ def create_config():
     if not os.path.exists('data'):
         os.makedirs('data')
 
-    print("config.json not found. Creating it now"
-    with open('data/config.json','w') as f:
-        f.write('{}')
+    if not os.path.exists('data/config.json'):
+        print("config.json not found. Creating it now")
+        with open('data/config.json','w') as f:
+            f.write('{}')
 
-    data = {}
+        data = dict()
+        data["channel_id"] = input("What channel id would you like to use? ")
 
-    data["channel_id"] = input("What channel id would you like to use? ")
-
-    with open('data/config.json','w') as f:
-        f.write(json.dumps(data))
+        with open('data/config.json','w') as f:
+            f.write(json.dumps(data))
 
 #creates .env file if you don't have one, and fills in the token
 def create_dotenv():
@@ -30,10 +31,18 @@ def create_dotenv():
         with open('.env','w') as env:
             env.write(f'TOKEN="{token}"')
 
+def create_db():
+    if not os.path.exists('data/database.db'):
+        print("database.db not found. Creating it now")
+        con = sqlite3.connect('data/database.db')
+        cur = con.cursor()
+        cur.execute('CREATE TABLE IF NOT EXISTS leaderboard(user_id, time, category)')
+
 def setup():
     install_requirements()
     create_config()
     create_dotenv()
+    create_db()
 
     input("Setup finished. Press enter to exit.")
     exit()
